fix: winning_stores grade 컬럼 타입 수정 및 Checkstyle 중괄호 위반 수정
- V17 grade TINYINT → INT (Hibernate 스키마 검증 오류 수정)
- LatestRoundController.taxRate if 블록 중괄호 추가 (NeedBraces)

+++ src/main/resources/db/migration/V17__create_winning_stores.sql
@@ -1,7 +1,7 @@
 CREATE TABLE winning_stores (
     id         BIGINT       NOT NULL AUTO_INCREMENT,
     round      INT          NOT NULL,
-    grade      TINYINT      NOT NULL,
+    grade      INT          NOT NULL,
     name       VARCHAR(200) NOT NULL,
     address    VARCHAR(300),
     win_count  INT          NOT NULL DEFAULT 1,
